+++ src/storage/txn/mod.rs
@@ -239,7 +239,7 @@ pub mod tests {
     pub use actions::acquire_pessimistic_lock::tests::{
         must_err as must_acquire_pessimistic_lock_err,
         must_err_return_value as must_acquire_pessimistic_lock_return_value_err,
-        must_succeed as must_acquire_pessimistic_lock,
+        must_pessimistic_locked, must_succeed as must_acquire_pessimistic_lock,
         must_succeed_for_large_txn as must_acquire_pessimistic_lock_for_large_txn,
         must_succeed_impl as must_acquire_pessimistic_lock_impl,
         must_succeed_return_value as must_acquire_pessimistic_lock_return_value,
